refactor: Migrate type imports from 'stores/types' to 'interfaces/interfaces' and remove unused types file
