+++ src/views/ClassManagementView.vue
@@ -5,7 +5,7 @@ import { defineAsyncComponent } from 'vue'
 import { useStudentsStore } from '../stores/studentsStore'
 import { useSectionsStore } from '../stores/sectionsStore'
 import { useClassesStore } from '../stores/coursesStore'
-import type { StudentProfile, StudentViewModel, YearLevel } from '../stores/types'
+import type { StudentProfile, StudentViewModel, YearLevel } from '../interfaces/interfaces'
 const Header = defineAsyncComponent(() => import('../components/Header.vue'))
 
 const route = useRoute()
